Lots of work on codecs:
New unit tests for codecs that have standard test data (with scripts to
regenerate the data).
Root classes are no longer allowed to have a type id. The reason for
this is because that ID would have to be in the global namespace and
would conflict with ids in other files.
More work on serializing and hashing enumeration types.
Move to deprecate extensions in favor of custom options.
Make sure struct / enum descriptors have their file property filled in.
Implemented 'bytes' type for text codecs (both cpp and Python).
Text encoders now handle character escape sequences in strings.
Various bug fixes in encoders.